Add Defaulted Implementation-Version Manifest Parsing
The `TypedManifest` class only knows about the Implementation-Version
header. The manifest is parsed using try-with-resources, which disposes
the resource safely after the try block completes or throws. This class
will return empty String for the following scenarios:

- file not found at path
- file found at path is not a manifest
- Implementation-Version key missing
- Implementation-Version key has no value

The `samples` module demonstrates how to include this library in a
project and start using it to print out Implementation-Version info.