Add Ability to Specify Default Manifest
An empty string response to a --version flag may not be the most helpful
result in all situations. For example, if I run a main class from an
IDE, I won't have an Implementation-Version in my MANIFEST.MF. Rather
than simply showing my version is "", I may want to show a message like
"Version Not Found".